requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Environment & Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic[email]

# Email Service
resend==2.4.0

# YAML parsing
pyyaml==6.0.2

# HTTP Client (for Medusa API)
httpx==0.27.2

# Jinja2 Templates
jinja2==3.1.4

# Optional: AI Enhancement (Phase 5)
# anthropic==0.39.0

